chore(b-e): Rebuild class path, ignore DS_Store
