+++ .vscode/launch.json
@@ -5,8 +5,8 @@
 			"name": "Spring Boot-DemoApplication<demo>",
 			"request": "launch",
 			"cwd": "${workspaceFolder}",
-			"mainClass": "com.example.demo.DemoApplication",
-			"projectName": "demo",
+			"mainClass": "app.web.lets_beat_google.DemoApplication",
+			"projectName": "lets_beat_google",
 			"args": "",
 			"envFile": "${workspaceFolder}/.env"
 		}
