--- Spring_Project/src/main/java/com/example/demo/Model/Keyword.java
@@ -1,10 +1,0 @@
-package com.example.demo.model;
-
-import lombok.Data;
-
-@Data
-public class Keyword {
-
-    private String text;
-    private int weight;
-}
