+++ backend/src/main/java/app/web/lets_beat_google/DemoApplication.java
@@ -1,0 +1,14 @@
+package app.web.lets_beat_google;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DemoApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(DemoApplication.class, args);
+
+  }
+
+}
