+++ backend/src/main/java/app/web/lets_beat_google/model/Keyword.java
@@ -1,0 +1,10 @@
+package app.web.lets_beat_google.model;
+
+import lombok.Data;
+
+@Data
+public class Keyword {
+
+  private String text;
+  private int weight;
+}
