+++ backend/src/main/java/app/web/lets_beat_google/service/GoogleSearchService.java
@@ -1,0 +1,16 @@
+package app.web.lets_beat_google.service;
+
+import java.util.ArrayList;
+
+import app.web.lets_beat_google.model.Website;
+
+public interface GoogleSearchService {
+
+  /**
+   * Get the search result of Google with the input query.
+   * 
+   * @param query
+   * @return the arraylist contains the search result
+   */
+  ArrayList<Website> getSearchResult(String query);
+}
