+++ backend/src/main/java/app/web/lets_beat_google/service/RelateWordService.java
@@ -1,0 +1,7 @@
+package app.web.lets_beat_google.service;
+
+import org.springframework.stereotype.Service;
+
+public interface RelateWordService {
+
+}
