+++ backend/src/main/java/app/web/lets_beat_google/service/impl/RelateWordServiceImpl.java
@@ -1,0 +1,7 @@
+package app.web.lets_beat_google.service.impl;
+
+import app.web.lets_beat_google.service.RelateWordService;
+
+public class RelateWordServiceImpl implements RelateWordService {
+
+}
