+++ backend/src/test/java/web/app/lets_beat_google/DemoApplicationTests.java
@@ -1,0 +1,13 @@
+package web.app.lets_beat_google;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DemoApplicationTests {
+
+  @Test
+  void contextLoads() {
+  }
+
+}
