+++ frontend/.gitignore
@@ -6,3 +6,4 @@ node_modules
 .output
 .env
 dist
+.DS_Store
